ESS: NeXus file writing enhancements and fixes
This fixes DM-1898
Adds a CacheStream element for streaming NICOS Kafka Cache data
Minor python3 compatability fixes

Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/22950
Tested-by: JenkinsCodeReview <bjoern_pedersen@frm2.tum.de>

+++ nicos_ess/devices/datasinks/imagesink/kafka.py
@@ -72,7 +72,7 @@ class ImageKafkaDataSinkHandler(DataSinkHandler):
         if quality not in [INTERMEDIATE, FINAL]:
             return
 
-        timestamp = long(currenttime() * 1E9)
+        timestamp = currenttime() * 1E9
         _, arrays = results[self.detector.name]
         for desc, array in zip(self.detector.arrayInfo(), arrays):
             self._sendArray(timestamp, desc, array)
